requirements.txt
altgraph==0.17.4
chardet==5.2.0
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.11
pywin32-ctypes==0.2.2
reportlab==4.0.8
